Move tests from `CardReaderPaymentViewModelTest` to `CardReaderPaymentController`
