+++ WooCommerce/src/test/kotlin/com/woocommerce/android/ui/payments/cardreader/CardReaderPaymentViewModelTest.kt
@@ -93,9 +93,6 @@ import com.woocommerce.android.ui.payments.receipt.PaymentReceiptShare
 import com.woocommerce.android.ui.payments.tracking.CardReaderTrackingInfoKeeper
 import com.woocommerce.android.ui.payments.tracking.PaymentsFlowTracker
 import com.woocommerce.android.util.CurrencyFormatter
-import com.woocommerce.android.util.PrintHtmlHelper.PrintJobResult.CANCELLED
-import com.woocommerce.android.util.PrintHtmlHelper.PrintJobResult.FAILED
-import com.woocommerce.android.util.PrintHtmlHelper.PrintJobResult.STARTED
 import com.woocommerce.android.util.captureValues
 import com.woocommerce.android.viewmodel.BaseUnitTest
 import com.woocommerce.android.viewmodel.MultiLiveEvent.Event
@@ -1827,213 +1824,7 @@ class CardReaderPaymentViewModelTest : BaseUnitTest() {
             assertThat((viewModel.event.value as PrintReceipt).receiptUrl).isEqualTo(receiptUrl)
             assertThat((viewModel.event.value as PrintReceipt).documentName).isEqualTo("receipt-order-1")
         }
-
-    @Test
-    fun `given external reader and receipt fetching and sharing success, when user clicks on send receipt button, then PlayChaChing emitted`() =
-        testBlocking {
-            whenever(cardReaderManager.collectPayment(any())).thenAnswer {
-                flow { emit(PaymentCompleted("url")) }
-            }
-            whenever(paymentReceiptShare("test url", 1L)).thenReturn(
-                PaymentReceiptShare.ReceiptShareResult.Success
-            )
-            viewModel.start()
-
-            (viewModel.viewStateData.value as ExternalReaderPaymentSuccessfulState).onSecondaryActionClicked.invoke()
-
-            assertThat(viewModel.event.value).isEqualTo(PlayChaChing)
-        }
-
-    @Test
-    fun `given built in reader and receipt fetching  and sharing success, when user clicks on send receipt button, then PlayChaChing emitted`() =
-        testBlocking {
-            whenever(cardReaderManager.collectPayment(any())).thenAnswer {
-                flow { emit(PaymentCompleted("")) }
-            }
-
-            initViewModel(BUILT_IN)
-            viewModel.start()
-
-            (viewModel.viewStateData.value as BuiltInReaderPaymentSuccessfulState).onSecondaryActionClicked.invoke()
-
-            assertThat(viewModel.event.value).isEqualTo(PlayChaChing)
-        }
-
-    @Test
-    fun `given receipt fetching success and receipt file not created, when user clicks on send receipt button, then ShowSnackbar emitted`() =
-        testBlocking {
-            whenever(cardReaderManager.collectPayment(any())).thenAnswer {
-                flow { emit(PaymentCompleted("url")) }
-            }
-            whenever(paymentReceiptShare("test url", 1L)).thenReturn(
-                PaymentReceiptShare.ReceiptShareResult.Error.FileCreation
-            )
-            viewModel.start()
-
-            (viewModel.viewStateData.value as ExternalReaderPaymentSuccessfulState).onSecondaryActionClicked.invoke()
-
-            assertThat((viewModel.event.value as ShowSnackbar).message).isEqualTo(
-                R.string.card_reader_payment_receipt_can_not_be_stored
-            )
-            verify(tracker).trackPaymentsReceiptSharingFailed(PaymentReceiptShare.ReceiptShareResult.Error.FileCreation)
-        }
-
-    @Test
-    fun `given receipt fetching success and receipt file not downloaded, when user clicks on send receipt button, then ShowSnackbar emitted`() =
-        testBlocking {
-            whenever(cardReaderManager.collectPayment(any())).thenAnswer {
-                flow { emit(PaymentCompleted("url")) }
-            }
-            whenever(paymentReceiptShare("test url", 1L)).thenReturn(
-                PaymentReceiptShare.ReceiptShareResult.Error.FileDownload
-            )
-            viewModel.start()
-
-            (viewModel.viewStateData.value as ExternalReaderPaymentSuccessfulState).onSecondaryActionClicked.invoke()
-
-            assertThat((viewModel.event.value as ShowSnackbar).message).isEqualTo(
-                R.string.card_reader_payment_receipt_can_not_be_downloaded
-            )
-            verify(tracker).trackPaymentsReceiptSharingFailed(PaymentReceiptShare.ReceiptShareResult.Error.FileDownload)
-        }
-
-    @Test
-    fun `given receipt fetching success and receipt file not shared, when user clicks on send receipt button, then ShowSnackbar emitted`() =
-        testBlocking {
-            whenever(cardReaderManager.collectPayment(any())).thenAnswer {
-                flow { emit(PaymentCompleted("url")) }
-            }
-            val sharing = PaymentReceiptShare.ReceiptShareResult.Error.Sharing(Exception())
-            whenever(paymentReceiptShare("test url", 1L)).thenReturn(sharing)
-            viewModel.start()
-
-            (viewModel.viewStateData.value as ExternalReaderPaymentSuccessfulState).onSecondaryActionClicked.invoke()
-
-            assertThat((viewModel.event.value as ShowSnackbar).message).isEqualTo(
-                R.string.card_reader_payment_email_client_not_found
-            )
-            verify(tracker).trackPaymentsReceiptSharingFailed(sharing)
-        }
-
-    @Test
-    fun `given external reader and receipt fetching fails, when user clicks on send receipt button, then ShowSnackabar event emitted`() =
-        testBlocking {
-            whenever(cardReaderManager.collectPayment(any())).thenAnswer {
-                flow { emit(PaymentCompleted("")) }
-            }
-            whenever(paymentReceiptHelper.getReceiptUrl(any())).thenReturn(Result.failure(Exception()))
-            viewModel.start()
-
-            (viewModel.viewStateData.value as ExternalReaderPaymentSuccessfulState).onSecondaryActionClicked.invoke()
-
-            assertThat((viewModel.event.value as ShowSnackbar).message).isEqualTo(R.string.receipt_fetching_error)
-        }
-
-    @Test
-    fun `given built reader and receipt fetching fails, when user clicks on send receipt button, then ShowSnackabar event emitted`() =
-        testBlocking {
-            whenever(cardReaderManager.collectPayment(any())).thenAnswer {
-                flow { emit(PaymentCompleted("")) }
-            }
-            whenever(paymentReceiptHelper.getReceiptUrl(any())).thenReturn(Result.failure(Exception()))
-
-            initViewModel(BUILT_IN)
-            viewModel.start()
-
-            (viewModel.viewStateData.value as BuiltInReaderPaymentSuccessfulState).onSecondaryActionClicked.invoke()
-
-            assertThat((viewModel.event.value as ShowSnackbar).message).isEqualTo(R.string.receipt_fetching_error)
-        }
-
-    @Test
-    fun `given external reader, when user clicks on send receipt button, then event tracked`() =
-        testBlocking {
-            whenever(cardReaderManager.collectPayment(any())).thenAnswer {
-                flow { emit(PaymentCompleted("")) }
-            }
-            viewModel.start()
-
-            (viewModel.viewStateData.value as ExternalReaderPaymentSuccessfulState).onSecondaryActionClicked.invoke()
-
-            verify(tracker).trackEmailReceiptTapped()
-        }
-
-    @Test
-    fun `given built in reader, when user clicks on send receipt button, then event tracked`() =
-        testBlocking {
-            whenever(cardReaderManager.collectPayment(any())).thenAnswer {
-                flow { emit(PaymentCompleted("")) }
-            }
-
-            initViewModel(BUILT_IN)
-
-            viewModel.start()
-
-            (viewModel.viewStateData.value as BuiltInReaderPaymentSuccessfulState).onSecondaryActionClicked.invoke()
-
-            verify(tracker).trackEmailReceiptTapped()
-        }
-
-    @Test
-    fun `given billing email empty and external, when user clicks on save for later button, then Exit event emitted`() =
-        testBlocking {
-            whenever(cardReaderManager.collectPayment(any())).thenAnswer {
-                flow { emit(PaymentCompleted("")) }
-            }
-            viewModel.start()
-
-            (viewModel.viewStateData.value as ExternalReaderPaymentSuccessfulState).onTertiaryActionClicked.invoke()
-
-            assertThat(viewModel.event.value).isInstanceOf(Exit::class.java)
-        }
-
-    @Test
-    fun `given billing email built in and external, when user clicks on save for later button, then Exit event emitted`() =
-        testBlocking {
-            whenever(cardReaderManager.collectPayment(any())).thenAnswer {
-                flow { emit(PaymentCompleted("")) }
-            }
-
-            initViewModel(BUILT_IN)
-            viewModel.start()
-
-            (viewModel.viewStateData.value as BuiltInReaderPaymentSuccessfulState).onTertiaryActionClicked.invoke()
-
-            assertThat(viewModel.event.value).isInstanceOf(Exit::class.java)
-        }
-
-    @Test
-    fun `given billing email not empty and external, when user clicks on save for later button, then Exit event emitted`() =
-        testBlocking {
-            whenever(mockedAddress.email).thenReturn("nonemptyemail")
-            whenever(cardReaderManager.collectPayment(any())).thenAnswer {
-                flow { emit(PaymentCompleted("")) }
-            }
-            viewModel.start()
-
-            (viewModel.viewStateData.value as ExternalReaderPaymentSuccessfulReceiptSentAutomaticallyState)
-                .onTertiaryActionClicked.invoke()
-
-            assertThat(viewModel.event.value).isInstanceOf(Exit::class.java)
-        }
-
-    @Test
-    fun `given billing email not empty and built in, when user clicks on save for later button, then Exit event emitted`() =
-        testBlocking {
-            whenever(mockedAddress.email).thenReturn("nonemptyemail")
-            whenever(cardReaderManager.collectPayment(any())).thenAnswer {
-                flow { emit(PaymentCompleted("")) }
-            }
-
-            initViewModel(BUILT_IN)
-            viewModel.start()
-
-            (viewModel.viewStateData.value as BuiltInReaderPaymentSuccessfulReceiptSentAutomaticallyState)
-                .onTertiaryActionClicked.invoke()
-
-            assertThat(viewModel.event.value).isInstanceOf(Exit::class.java)
-        }
-
+    //
     @Test
     fun `given user presses back button, when re-fetching order, then ReFetchingOrderState shown`() =
         testBlocking {
